update graphViewMaker
- 두번째 페이지 함수 변경 및 내용 수정
